+++ ChangeLog.md
@@ -6,7 +6,7 @@
 
     Release date: 
 
-    - Now relies on pwix:modal (todo #1)
+    - Now relies on pwix:modal v 1.6.0 (todo #1)
     - Clicking on the backdrop no more close the dialogs (todo #3)
 
 ### 1.5.0